flask_replica/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
google-generativeai
a2wsgi
pytz
python-multipart
starlette
uvloop
flask
flask-cors
chromadb
flask-sqlalchemy
flask-login

